pyproject.toml
[project]
name = "autoresearch"
version = "0.2.0"
description = "Lightweight harness for AI-driven code experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
autoresearch = "autoresearch.__main__:main"
